+++ derive/src/object.rs
@@ -350,6 +350,7 @@ pub fn generate(
                         default_with,
                         validator,
                         visible,
+                        secret,
                         ..
                     },
                 ) in &args
@@ -392,6 +393,7 @@ pub fn generate(
                             default_value: #schema_default,
                             validator: #validator,
                             visible: #visible,
+                            is_secret: #secret,
                         });
                     });