+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.8.2] 2021-04-09
+
+- Now when the resolver returns the `Result` type, `E` can be all types that implement `async_graphql::Into<Error>`.
+
+## [2.8.1] 2021-04-08
 
 ### Fixed